Move locker into own directory.
