+++ lsif/src/worker/processors/convert.ts
@@ -2,12 +2,13 @@ import * as constants from '../../shared/constants'
 import * as fs from 'mz/fs'
 import * as path from 'path'
 import * as settings from '../settings'
+import { addTags, TracingContext } from '../../shared/tracing'
 import { convertLsif } from '../importer/importer'
 import { createSilentLogger } from '../../shared/logging'
 import { dbFilename } from '../../shared/paths'
-import { TracingContext, addTags } from '../../shared/tracing'
-import { XrepoDatabase, PostgresLocker } from '../../shared/xrepo/xrepo'
 import { Job } from 'bull'
+import { PostgresLocker } from '../../shared/locker/locker'
+import { XrepoDatabase } from '../../shared/xrepo/xrepo'
 
 /**
  * Create a job that takes a repository, commit, and filename containing the gzipped
